Allow use paths into sub-module of expanded lib
Now the binary can have use items such as
`use <crate-name>::internal::hello_world;`. However, it is not perfect
and only works for use paths starting with the <crate-name>, as well as
only working correctly from the global scope of the binary. If you use
such a path in a sub-module of the binary, you will get an incorrect
use path.